sdl: create wayland EGL window as SDL is not responsible for it anymore
We need to pass the EGLNativeWindowType handle to the renderer and it
needs to be responsible for creating the EGLSurface. With upstream SDL
this is not the case as it wants to maintain ownership over the windows
EGLSurface and doesn't allow us to get access to it. Instead of
refactoring our code we live with a hacked SDL for the moment.